Create seperate http api instance
