+++ src/echo-server.ts
@@ -2,6 +2,7 @@ import { HttpSubscriber, RedisSubscriber } from './subscribers';
 import { Channel } from './channels';
 import { Server } from './server';
 import { Log } from './log';
+import { HttpApi } from './api';
 
 const packageFile = require('../package.json');
 
@@ -70,6 +71,13 @@ export class EchoServer {
     private httpSub: HttpSubscriber;
 
     /**
+     * Http api instance.
+     *
+     * @type {HttpApi}
+     */
+    private httpApi: HttpApi;
+
+    /**
      * Create a new instance.
      */
     constructor() { }
@@ -104,7 +112,9 @@ export class EchoServer {
         return new Promise((resolve, reject) => {
             this.channel = new Channel(io, this.options);
             this.redisSub = new RedisSubscriber(this.options);
-            this.httpSub = new HttpSubscriber(io, this.channel, this.options, this.server.express);
+            this.httpSub = new HttpSubscriber(this.server.express);
+            this.httpApi = new HttpApi(io, this.channel, this.server.express);
+            this.httpApi.init();
 
             this.onConnect();
             this.listen().then(() => resolve());
